Refactor integer types and add Short/Long value types
- Change Integer from int64 to int as the default integer type
- Add Short (int16) and Long (int64) value types with full arithmetic support
- Update Numeric interface Neg() to return (Value, error)
- Rename OpLoadAttr to OpLoadMember for clarity
- Add Memberable interface to NamespaceValue and Stream
- Add Time value type and now() builtin function
- Update REPL formatting for new value types
- Fix various type conversions throughout codebase

+++ cmd/vega/cli/root.go
@@ -93,7 +93,7 @@ VFS-mounted storage backends (SQLite, S3, PostgreSQL, ephemeral, etc.)`,
 					return fmt.Errorf("runtime error: %w", err)
 				}
 
-				vm.SetGlobal("exitcode", value.NewInteger(int64(exitCode)))
+				vm.SetGlobal("exitcode", value.NewInteger(exitCode))
 				// If interactive is 'false' (default), close immediately to avoid running vega
 				if !interactive {
 					return fs.Shutdown(ctx)
